Add doc support and other CI related files
This change adds verious bits needed for CI and/or documentation
scraping as well as some README files.

The version in the CMake file was set to 0.1.0 in prep for activation
of CI jobs -- we want to be able to fiddle with those without actually
pushing a 1.x.x package to the repo. Once CI jobs are working we can
up the version to 1.0.0 to match the AP version.

Issue-ID: RIC-684

Signed-off-by: E. Scott Daniels <daniels@research.att.com>
Change-Id: Iac2c24a8e8f1c01af74e20c0bce9cad1c6bf6702